requirements.txt
python-dateutil==2.7.2
Django>=1.11,<2.1
django-swingtime>=0.9.0
-e git+https://github.com/deschler/django-modeltranslation.git#egg=django-modeltranslation
django-heroku
python3-indy
python3-wrapper-vcx
django-extensions
pygraphviz
pydotplus
djangorestframework
markdown
django-filter
django-background-tasks
